Added a new methods of controller and fixed services
Added a search by taken status for book, also, a search by name, email and phone for user. Moved all logic from ContractsController to ContractService. Removed unused method in the service, daoImpl and dao of Book. Worked out the error NoResultException in the UserDaoImpl. Override the method of save and delete user, so this way now we cannot create several users with the same email of phone number