Moved methods between services
In this commit, I refactored two methods in the ContractService class to the BookService class. These methods are related to the Book entity and their presence in the ContractService class was violating the Single Responsibility Principle. Moving them to the BookService class improves the code's adherence to SOLID principles. Also, fixed the circle dependents between BooksController and ContractService.