Create tests for part of controller and service
Added tests for the AuthorsController, implemented in AuthorsControllerTest, and for the ContractService. The tests for the service layer cover the methods findIsOverdue, findIsOverdueByUser, extendDateOfReturn, removeContractAfterReturned, createContract, setFieldsForBorrow, setContractFieldsOnBorrow, checkDateOfReturned, and checkAvailabilityForDeleteContract. The tests provide coverage for the functionality of these methods, ensuring their correct operation and that they return the expected results. Also, added a methods for AuthorsController, this method show the books of the author.